feat: Huggis 주석 추가

+++ src/main/java/young/refactoring/ch9/splitVariableExam/Haggis.java
@@ -1,15 +1,19 @@
 package young.refactoring.ch9.splitVariableExam;
 
 public class Haggis {
+
     public double distanceTravelled(Scenario scenario, int time) {
         double result;
-        int acc = scenario.primaryForce / scenario.mass; // 가속도 = 힘 * 질량
-        int primaryTime = Math.min(time, scenario.delay);
+        // 초기힘 적용
+        double acc = (double) scenario.primaryForce / scenario.mass; // 가속도 = 힘 * 질량
+        double primaryTime = Math.min(time, scenario.delay);
         result = 0.5 * acc * primaryTime * primaryTime; // 전파된 거리
-        int secondaryTime = time - scenario.delay;
+
+        // 두번째 힘 적용
+        double secondaryTime = time - scenario.delay;
         if (secondaryTime > 0) {
             // 두 번째 힘을 반영해 다시 계산
-            int primaryVelocity = acc * scenario.delay;
+            double primaryVelocity = acc * scenario.delay;
             acc = (scenario.primaryForce + scenario.secondaryForce) / scenario.mass;
             result += primaryVelocity * secondaryTime + 0.5 * acc * secondaryTime * secondaryTime;
         }
